### 2024-01-11 #6214 @editorialbot
[PRE REVIEW]: MousebreedeR: A novel software to assist in the design of breeding schema for complex genotypes of experimental organisms
**Submitting author:** @sportiellomike (<a href="http://orcid.org/0000-0003-1690-8702">Mike Sportiello</a>)
**Repository:** https://github.com/sportiellomike/mousebreedeR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
