### 2024-01-17 #6214 @Kevin-Mattheus-Moerman
@openjournals/dev it looks like out system did not detect the non-approved licence here. :point_up:
