### 2024-01-17 d758b57
Update LICENSE.md
MIT license per https://opensource.org/license/mit/.
