### 2024-01-17 #6214 @sportiellomike
@Kevin-Mattheus-Moerman my mistake. It has since been updated to the MIT license which is per https://opensource.org/license/mit/. Thank you.
