### 2024-01-29 #6214 @csoneson
Sure, I can take this on.
@sportiellomike - thanks for your submission. Before/while I start looking for reviewers, I have a few comments from a first look at your submission:

- I would suggest adding some more explanatory text to the vignette - at the moment it's all code chunks with some comments, I think it would be easier to read if there was more actual text helping the reader to understand the code and interpret the output.
- It is generally recommended not to depend directly on the `tidyverse` package (see https://www.tidyverse.org/blog/2018/06/tidyverse-not-for-packages/). Could you depend on individual components instead (also, do you really need it in Depends or could you move it to Imports)?
- The repository contains some files that look like they shouldn't be there (`deletethese`, `oldvigs`, some R scripts in the main folder), could you clean that up?
- I would strongly recommend adding unit tests that check that the package code generates the expected output. Note that this is also one of the points in the JOSS review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html).
- Please remove the installation instructions from the paper, the user will get this from the  GitHub repository.
