+++ vignettes/Vignette.Rmd
@@ -8,7 +8,8 @@ vignette: >
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
-
+# Walkthrough
+First we start off by loading all of the necessary libraries.
 ```{r}
 library(mousebreedeR)
 library(dqrng)
@@ -22,7 +23,11 @@ library(ggpubr)
 `%!in%` <- Negate(`%in%`)
 ```
 
+The package takes a table of mouse genotypes and gives the user breeding strategy. The user supplies a data frame with a numeric number of copies of a certain allele. For example, for the allele A, the genotype AA would have 2 copies, Aa 1 copy, and aa 0 copies. The data frame supplied should use actual numbers, not phrases like "homozygous positive" or "AA," for example. 
 
+In the below code chunk, we can build an example data frame. The last argument `head(df)` will show a properly formatted input to the package, but most people will find it easier to read in a csv.
+
+YOU CAN REPLACE `df` WITH YOUR DATAFRAME OF INTEREST AND RUN THE BELOW CODE. MAKE SURE TO ALSO UPDATE THE `desiredvec` BELOW APPROPRIATELY.
 ```{r}
 # make example df
 mouse1 <- c(0, 0, 0, 1, 0, 0, 0, 1)
@@ -33,8 +38,15 @@ mouse5 <- c('M', 'M', 'M', 'M','F', 'F', 'F', 'F')
 df <- data.frame(mouse1, mouse2, mouse3, mouse4, mouse5)
 colnames(df) <- c(paste0('gene', 2:dim(df)[2] - 1), 'sex')
 head(df)
+# rename df
+
+# df<- SAVE YOUR DATAFRAME OF AVAILABLE MICE TO THE VARIABLE df. DON'T FORGET TO UPDATED desiredvec BELOW.
+exampleData<-df
 ```
 
+As you can see above, the format of the data frame should be as follows: each column is an allele (A or B or C, for example), each row is a mouse. Each cell should have the number of that allele in that mouse. For a mouse that is AABbCcdd, the row should have 2 in the first column, 1 in the second column, 1 in the third column, and 0 in the fourth column. The last column should be the sex of the animal: M for male, and F for female.
+
+In the next code chunk, we set the theme for the plots we make later on.
 ```{r}
 # set ggplothemes
 theme_set(
@@ -52,76 +64,72 @@ theme_set(
     panel.grid.major = element_line(color = 'light grey')
   )
 )
-
-# rename df
-exampleData<-df
-head(exampleData) # look at the top few rows
 ```
 
+First let's look at what happens when each mouse undergoes meiosis and creates gametes. Then, we can look at the possible gametes produced by the mice from the data frame you originally created.
 ```{r}
-# first let's look at what happens when each mouse undergoes meiosis
 meiosisoutput<-engageinmeiosis(exampleData)
-head(meiosisoutput) # as you can see: some of these have an impossible number of alleles (0.5 copies). The next function will fix that by generating gametes that have either 0 or 1 copies of the allele like in real meiosis.
+# head(meiosisoutput) # as you can see: some of these have an impossible number of alleles (0.5 copies). The next function will fix that by generating gametes that have either 0 or 1 copies of the allele like in real meiosis.
 
 compilegametesoutput<-compilegametes(meiosisoutput)
 head(compilegametesoutput)
 ```
 
+Let's look specifically at the genotypes of sperm, and then eggs.
 ```{r}
 # now we can separate them into sperm and eggs
 spermandeggs(x = compilegametesoutput,sex = 'sex')
 head(sperm)
 head(eggs)
 ```
-
[…]
