+++ mouseschmerplane.R
@@ -1,609 +1,0 @@
-#mouseschemer
-library(dqrng)
-library(dplyr)
-library(gtools)
-library(tidyverse)
-library(ggplot2)
-library(reshape2)
-library(viridis)
-library(ggpubr)
-`%!in%` <- Negate(`%in%`)
-# make example df
-mouse1 <- c(2, 1, 0, 2, 2, 1, 1, 1, 0)
-mouse2 <- c(0, 1, 0, 2, 1, 0, 0, 2, 1)
-mouse3 <- c(2, 1, 2, 0, 2, 1, 1, 1, 0)
-mouse4 <- c(1, 1, 0, 2, 2, 2, 2, 0, 2)
-mouse5 <- c('M', 'F', 'M', 'M', 'F', 'M', 'F', 'M', 'F')
-df <- data.frame(mouse1, mouse2, mouse3, mouse4, mouse5)
-
-str(df)
-df
-colnames(df) <- c(paste0('gene', 2:dim(df)[2] - 1), 'sex')
-df
-str(df)
-
-# encode definitions
-homoneg <- c(
-  0,
-  NA,
-  'homoneg',
-  'neg',
-  'doubleneg',
-  'homoneg',
-  'homonegative',
-  'homozygousneg',
-  'homozygous negative',
-  'aa','bb','cc','dd','ee','ff','gg','hh','ii','jj','kk',
-  'll','mm','nn','oo','pp','qq','rr','ss','tt','uu','vv','ww','xx','yy','zz'
-)
-
-homopos <- c(2,
-             'doublepos',
-             'homopos',
-             'homopositive',
-             'homozygouspos',
-             'homozygous positive',
-             'AA','BB','CC','DD','EE','FF','GG','HH','II','JJ','KK',
-             'LL','MM','NN','OO','PP','QQ','RR','SS','TT','UU','VV','WW','XX','YY','ZZ')
-het <- c(1,
-         'pos',
-         'Aa','Bb','Cc','Dd','Ee','Ff','Gg','Hh','Ii','Jj','Kk',
-         'Ll','Mm','Nn','Oo','Pp','Qq','Rr','Ss','Tt','Uu','Vv','Ww','Xx','Yy','Zz')
-
-male <- c('M', 'Male', 'MALE')
-female <- c('F', 'Female', 'FEMALE')
-
-males <- subset(df, df$sex %in% male)
-females <- subset(df, df$sex %in% female)
-
-# set ggplothemes
-theme_set(theme(axis.text = element_text(size=8),
-                axis.title = element_text(size=8),
-                strip.text = element_text(size=8),
-                axis.text.x = element_text(angle=90),
-                legend.position = 'none',
-                # strip.text.x = element_text(size = 8,margin = margin(.1, 0, .1, 0, "cm")),
-                legend.text = element_text(size=7),
-                legend.title = element_text(size=7),
-                # legend.position = 'bottom',panel.border=element_blank(),
-                panel.background = element_blank(),
-                panel.grid.major = element_line(color='light grey')
-))
-# write functions
-addmouseID <- function(x) {
-  x$mouseID <- paste0('mouse', 1:dim(x)[1])
-  return(x)
-}
-
-engageinmeiosis <- function(x) {
-  x <- addmouseID(x)
-  colnumerics <- as.character(lapply(x, FUN = is.numeric))
-  indexcolnumerics <- which(colnumerics == T)
-  indexcolchar <- which(colnumerics == F)
-  gametegenotypes <- x[indexcolnumerics]
-  gametemetadata <- x[indexcolchar]
-
-  indexhomopos <- which(gametegenotypes == homopos)
-  indexhomoneg <- which(gametegenotypes == homoneg)
-  indexhet <- which(gametegenotypes == het)
-
[…]
