+++ oldvigs/vignette2.R
@@ -1,240 +1,0 @@
-#mousebreedeR
-library(dqrng)
-library(dplyr)
-library(gtools)
-library(tidyverse)
-library(ggplot2)
-library(reshape2)
-library(viridis)
-library(ggpubr)
-Rfunctionstoload<-list.files(path = './R/',full.names = T)
-lapply(Rfunctionstoload,source)
-`%!in%` <- Negate(`%in%`)
-# make example df
-mouse1 <- c(0, 0, 0, 1, 0, 0, 0, 1)
-mouse2 <- c(0, 0, 1, 0, 0, 0, 1, 0)
-mouse3 <- c(0, 1, 0, 0, 0, 1, 0, 0)
-mouse4 <- c(1, 0, 0, 0, 1, 0, 0, 0)
-mouse5 <- c('M', 'M', 'M', 'M','F', 'F', 'F', 'F')
-df <- data.frame(mouse1, mouse2, mouse3, mouse4, mouse5)
-colnames(df) <- c(paste0('gene', 2:dim(df)[2] - 1), 'sex')
-head(df)
-#
-
-
-# set ggplothemes
-theme_set(
-  theme(
-    axis.text = element_text(size = 8),
-    axis.title = element_text(size = 8),
-    strip.text = element_text(size = 8),
-    axis.text.x = element_text(angle = 90),
-    legend.position = 'none',
-    # strip.text.x = element_text(size = 8,margin = margin(.1, 0, .1, 0, "cm")),
-    legend.text = element_text(size = 7),
-    legend.title = element_text(size = 7),
-    # legend.position = 'bottom',panel.border=element_blank(),
-    panel.background = element_blank(),
-    panel.grid.major = element_line(color = 'light grey')
-  )
-)
-
-# rename df
-exampleData<-df
-head(exampleData) # look at the top few rows
-
-# first let's look at what happens when each mouse undergoes meiosis
-meiosisoutput<-engageinmeiosis(exampleData)
-head(meiosisoutput) # as you can see: some of these have an impossible number of alleles (0.5 copies). The next function will fix that by generating gametes that have either 0 or 1 copies of the allele like in real meiosis.
-
-compilegametesousput<-compilegametes(meiosisoutput)
-head(compilegametesousput)
-
-# now we can separate them into sperm and eggs
-spermandeggs(x = compilegametesousput,sex = 'sex')
-head(sperm)
-head(eggs)
-
-# now we can actually have sex
-fertilizeoutput<-fertilize(malegametes = sperm,
-          femalegametes = eggs)
-
-head(fertilizeoutput) # as you can see, we have genotypes, and we recorded who the mom and dad were. We also have a term called momdad that essentially saves that crossing (ie it pastes together the mom and the dad)
-
-# next we have two ways of summarizing the outcome of the crossing.
-# the first records the distribution of each genotype per momdad (ie per cross) and per gene
-summarizefertilizationoutput<-summarizefertilization(fertilizeoutput)
-head(summarizefertilizationoutput)
-
-# the next way to summarize the data is probably more useful, and that's to generate the probability that each potential pup is actually born. Included in this summary are the probabilities a pup of that genotype is not born given a litter size from 1-10 pups. For example, "notthatgenotypefivepup" column gives the probability that, if 5 pups are born, what is the probability that none of them are that genotype? You can subtract those columns from 1 to get the probability that at least one pup is that genotype.
-summarizepotentialpupoutput<-summarizepotentialpups(fertilizeoutput)
-head(summarizepotentialpupoutput)
-engageinmeiosis(summarizepotentialpupoutput)
-
-
-# let's set a goal
-head(fertilizeoutput)
[…]
