### 2024-01-30 1dd8dad
Updated descriptions of functionis
