### 2024-01-30 9903c47
Updated description
moved tidyverse from depends to imports, per suggestion of journal
