+++ DESCRIPTION
@@ -9,10 +9,10 @@ License: CC BY 4.0
 Encoding: UTF-8
 LazyData: true
 Depends: 
-    R (>= 2.10),
-    tidyverse
+    R (>= 2.10)
 RoxygenNote: 7.2.1
 Imports: 
+    tidyverse,
     dplyr,
     dqrng,
     ggplot2,
