+++ vignettes/Vignette.Rmd
@@ -30,12 +30,12 @@ In the below code chunk, we can build an example data frame. The last argument `
 YOU CAN REPLACE `df` WITH YOUR DATAFRAME OF INTEREST AND RUN THE BELOW CODE. MAKE SURE TO ALSO UPDATE THE `desiredvec` BELOW APPROPRIATELY.
 ```{r}
 # make example df
-mouse1 <- c(0, 0, 0, 1, 0, 0, 0, 1)
-mouse2 <- c(0, 0, 1, 0, 0, 0, 1, 0)
-mouse3 <- c(0, 1, 0, 0, 0, 1, 0, 0)
-mouse4 <- c(1, 0, 0, 0, 1, 0, 0, 0)
-mouse5 <- c('M', 'M', 'M', 'M','F', 'F', 'F', 'F')
-df <- data.frame(mouse1, mouse2, mouse3, mouse4, mouse5)
+gene1 <- c(0, 0, 0, 1, 0, 0, 0, 1)
+gene2 <- c(0, 0, 1, 0, 0, 0, 1, 0)
+gene3 <- c(0, 1, 0, 0, 0, 1, 0, 0)
+gene4 <- c(1, 0, 0, 0, 1, 0, 0, 0)
+gene5 <- c('M', 'M', 'M', 'M','F', 'F', 'F', 'F')
+df <- data.frame(gene1, gene2, gene3, gene4, gene5)
 colnames(df) <- c(paste0('gene', 2:dim(df)[2] - 1), 'sex')
 head(df)
 
@@ -101,7 +101,6 @@ The next way to summarize the data is probably more useful, and that's to genera
 ```{r}
 summarizepotentialpupoutput<-summarizepotentialpups(fertilizeoutput)
 head(summarizepotentialpupoutput)
-engageinmeiosis(summarizepotentialpupoutput)
 ```
 
 Let's say you have a specific goal of a genotype you want to create through breeding. Let's say we want homozygous positive mouse at 4 loci: AABBCCDD. Code this by making a vector of your desired genotype (let's call it your `desiredvec`). This desiredvec of genotypes needs to be the same length as the number of gene columns you have in your fertilizeoutput (which will be the same as what you supplied in the beginning to this walkthrough). Use the following words to make your desiredvec: homopos, het, homoneg. 
