+++ tests/testthat/test-compilegametes.R
@@ -1,4 +1,14 @@
 test_that("gametes compile", {
+  library(mousebreedeR)
+  library(dqrng)
+  library(dplyr)
+  library(gtools)
+  library(tidyverse)
+  library(ggplot2)
+  library(reshape2)
+  library(viridis)
+  library(ggpubr)
+  `%!in%` <- Negate(`%in%`)
   # first make the df
   # make example df
   gene1 <- c(0, 0, 0, 1, 0, 0, 0, 1)
