+++ R/notin.R
@@ -1,0 +1,11 @@
+#' addpoints
+#'
+#' @description Negates %in%
+#'
+#' @return negated %in%
+#' @export
+#'
+#' @examples
+#' people %!in% distress
+
+`%!in%` <- Negate(`%in%`)
