+++ man/grapes-not-in-grapes.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/notin.R
+\name{\%!in\%}
+\alias{\%!in\%}
+\title{addpoints}
+\usage{
+... \%!in\% NA
+}
+\value{
+negated %in%
+}
+\description{
+Negates %in%
+}
+\examples{
+people \%!in\% distress
+}
