+++ DESCRIPTION
@@ -10,7 +10,7 @@ Encoding: UTF-8
 LazyData: true
 Depends: 
     R (>= 2.10)
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.3
 Imports: 
     tidyverse,
     dplyr,
