### 2024-01-30 #6214 @sportiellomike
Thank you @csoneson for your helpful comments!

- Vignette edited!
- Changed dependency for tidyverse!
- Cleaned up folder!
- Added tests!
- Removed installation section from paper and am about to regenerate file for JOSS!

Thanks again!
