### 2024-02-12 #6214 @csoneson
👋🏻 @yjunechoe, @Balexandre23 (or @JoanyMarino, @btmonier, @mkerin) - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> MousebreedeR: A novel software to assist in the design of breeding schema for complex genotypes of experimental organisms

https://github.com/openjournals/joss-reviews/issues/6214

Thanks a lot in advance!
