### 2024-02-12 #6214 @yjunechoe
I am happy to review, though please note that I do not have domain expertise in this area. I will limit my comments to the quality of the code.

I'd also like to note a couple immediate concerns from a cursory look at the codebase:

- The package (still) imports the entire `tidyverse` suite of packages. Functions from external dependencies (e.g., `bind_rows()` from `dplyr`) are not appropriately imported or namespaced where they are used within the package.
- `NAMESPACE` exports everything, and does not discriminate between user facing and internal functions
- Code in `@examples` are not minimally reproducible examples
- Code in `/tests` contain unnecessary copy-pasted code and do not isolate unit tests
- Documentation files in `/man` contain what I presume to be leftovers from previous versions of the package (e.g., `hello.Rd`)

Speaking for myself as a potential reviewer, I don't think this package is quite ready for review. I would like for at least these issues to be resolved before I personally dedicate more time to reviewing the actual functionalities of the package. In sum, I conditionally accept to be a reviewer.
