### 2024-02-13 #6214 @sportiellomike
Thank you @yjunechoe for your time,

- I think I am misunderstanding something here, since I was asked to move it to imports which I did (see above thread). Could you help me understand what you mean?
- I used the roxygen2 (https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html) to help create the package, as described in most R packages, which created a default NAMESPACE. Is there utility in editing it?
- The code in @examples follows the vignette's naming conventions to walk the user through the process. Should I make sure the @example has extra lines in it analogous to the vignette to make reproducible?
- The tests I created attempted to make sure the package's main arguments worked successfully. Would it be better if I made sure the function tested got a specific value each time that I have preset?
- hello.Rd should have been deleted, and that was my mistake for missing that. It is now corrected.
