### 2024-02-13 #6214 @yjunechoe
Thank you for the quick response addressing my comments - I'll answer your questions in turn:

1)

> - I think I am misunderstanding something here, since I was asked to move it to imports which I did (see above thread). Could you help me understand what you mean?

It is indeed an improvement to move a dependency from Depends to Imports, but this still misses the bigger issue about depending on the entire tidyverse metapackage (Imports still creates a dependency - for more on the technical details, see https://r-pkgs.org/dependencies-in-practice.html#sec-dependencies-in-imports). The package should specifically Import only the packages necessary for it to function - and in fact, since the package already Imports dplyr and ggplot2, that may be sufficient on their own.

2)

> - I used the roxygen2 (https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html) to help create the package, as described in most R packages, which created a default NAMESPACE. Is there utility in editing it?

I would highly advise using `@export` tags for each function that you plan to export, but perhaps this not so critical for a package of this (small) scale. I will take this back for the moment and evaluate the necessity of this in the actual review.

3)

> - The code in https://github.com/examples follows the vignette's naming conventions to walk the user through the process. Should I make sure the @example has extra lines in it analogous to the vignette to make reproducible?

I mean that simple, one-liner examples with no setup [like this](https://github.com/sportiellomike/mousebreedeR/blob/5029776af30ffcec1c2bba1f94a3a51f29e76c82/R/compilegametes.R#L9C1-L10C41):

```r
#' @examples
#' compilegametes(engageinmeiosisoutput)
```

cannot be ran by the user after installing and loading the package (what is the `engageinmeiosisoutput` here and where is it defined?). This makes it very difficult to evaluate the functionalities of the functions in the package as a reviewer, and constitutes incomplete documentation of a function's behavior. Please ensure that all code in `@examples` includes the necessary setup such that they can be ran as-is, once the package has been loaded.

4)

> - The tests I created attempted to make sure the package's main arguments worked successfully. Would it be better if I made sure the function tested got a specific value each time that I have preset?

As I have not yet fully evaluated the scope of the package's features, I will also take this back for the moment. But I would advise you to consult existing materials on writing unit tests in R to increase the number and kind of unit tests (e.g., https://r-pkgs.org/testing-design.html), to receive a better evaluation of the test suite in the review (tests should not only cover correctness, but also edge cases, the signaling of appropriate messages and warnings, etc.). In any case, I can say more on what I expect to see from tests in the actual review.

---

Please address (1) and (3) and I am happy to formally review the package.
