### 2024-02-18 04c7346
created data that gets loaded with package to make minimally reproducible examples work
