+++ vignettes/Vignette.Rmd
@@ -15,7 +15,7 @@ library(mousebreedeR)
 library(dqrng)
 library(dplyr)
 library(gtools)
-# library(tidyverse)
+library(tidyverse)
 library(ggplot2)
 library(reshape2)
 library(viridis)
