+++ DESCRIPTION
@@ -12,6 +12,7 @@ Depends:
     R (>= 2.10)
 RoxygenNote: 7.2.3
 Imports: 
+    tidyverse,
     dplyr,
     dqrng,
     ggplot2,
