+++ data/data.R
@@ -1,160 +1,160 @@
 ## ----message=FALSE, warning=FALSE---------------------------------------------------------------------------------------------------
-library(mousebreedeR)
-library(dqrng)
-library(dplyr)
-library(gtools)
-# library(tidyverse)
-library(ggplot2)
-library(reshape2)
-library(viridis)
-library(ggpubr)
-`%!in%` <- Negate(`%in%`)
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-# make example df
-gene1 <- c(0, 0, 0, 1, 0, 0, 0, 1)
-gene2 <- c(0, 0, 1, 0, 0, 0, 1, 0)
-gene3 <- c(0, 1, 0, 0, 0, 1, 0, 0)
-gene4 <- c(1, 0, 0, 0, 1, 0, 0, 0)
-gene5 <- c('M', 'M', 'M', 'M','F', 'F', 'F', 'F')
-df <- data.frame(gene1, gene2, gene3, gene4, gene5)
-colnames(df) <- c(paste0('gene', 2:dim(df)[2] - 1), 'sex')
-head(df)
-
-# df<- SAVE YOUR DATAFRAME OF AVAILABLE MICE TO THE VARIABLE df. DON'T FORGET TO UPDATED desiredvec BELOW.
-
-# rename df
-exampleData<-df
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-# set ggplothemes
-theme_set(
-  theme(
-    axis.text = element_text(size = 8),
-    axis.title = element_text(size = 8),
-    strip.text = element_text(size = 8),
-    axis.text.x = element_text(angle = 90),
-    legend.position = 'none',
-    # strip.text.x = element_text(size = 8,margin = margin(.1, 0, .1, 0, "cm")),
-    legend.text = element_text(size = 7),
-    legend.title = element_text(size = 7),
-    # legend.position = 'bottom',panel.border=element_blank(),
-    panel.background = element_blank(),
-    panel.grid.major = element_line(color = 'light grey')
-  )
-)
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-meiosisoutput<-engageinmeiosis(exampleData)
-# head(meiosisoutput) # as you can see: some of these have an impossible number of alleles (0.5 copies). The next function will fix that by generating gametes that have either 0 or 1 copies of the allele like in real meiosis.
-
-compilegametesoutput<-compilegametes(meiosisoutput)
-head(compilegametesoutput)
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-# now we can separate them into sperm and eggs
-spermandeggs(x = compilegametesoutput,sex = 'sex')
-head(sperm)
-head(eggs)
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-fertilizeoutput<-fertilize(malegametes = sperm,
-          femalegametes = eggs)
-
-head(fertilizeoutput)
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-summarizefertilizationoutput<-summarizefertilization(fertilizeoutput)
-head(summarizefertilizationoutput)
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-summarizepotentialpupoutput<-summarizepotentialpups(fertilizeoutput)
-head(summarizepotentialpupoutput)
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
-desiredvec<-c('homopos', 'homopos', 'homopos', 'homopos')  # UPDATE THIS desiredvec APPROPRIATELY FOR YOUR NEEDS
-
-
-## -----------------------------------------------------------------------------------------------------------------------------------
[…]
