+++ data/data.R
@@ -159,32 +159,81 @@
 
 ## -----------------------------------------------------------------------------------------------------------------------------------
 
-# Here we can save all the data for minimally reproducible examples for documentation purposes.
-#
-# usethis::use_data(exampleData)
-#
-# usethis::use_data(meiosisoutput)
-#
-# usethis::use_data(compilegametesoutput)
-#
-# usethis::use_data(sperm)
-#
-# usethis::use_data(eggs)
-#
-# usethis::use_data(fertilizeoutput)
-#
-# usethis::use_data(summarizefertilizationoutput)
-#
-# usethis::use_data(summarizepotentialpupoutput)
-#
-# usethis::use_data(desiredvec)
-#
-# usethis::use_data(canwegetalltheallelesfromonecross)
-#
-# usethis::use_data(pointsperpupoutput)
-#
-# usethis::use_data(whichpairstobreed)
-#
-#
+Here we can save all the data for minimally reproducible examples for documentation purposes.
 
+usethis::use_data(exampleData)
+
+usethis::use_data(meiosisoutput)
+
+usethis::use_data(compilegametesoutput)
+
+usethis::use_data(sperm)
+
+usethis::use_data(eggs)
+
+usethis::use_data(fertilizeoutput)
+
+usethis::use_data(summarizefertilizationoutput)
+
+usethis::use_data(summarizepotentialpupoutput)
+
+usethis::use_data(desiredvec)
+
+usethis::use_data(canwegetalltheallelesfromonecross)
+
+usethis::use_data(pointsperpupoutput)
+
+usethis::use_data(whichpairstobreed)
+
+
+##########
+
+exampleexampleData<-exampleData
+
+examplemeiosisoutput<-meiosisoutput
+
+examplecompilegametesoutput<-compilegametesoutput
+
+examplesperm<-sperm
+
+exampleeggs<-eggs
+
+examplefertilizeoutput<-fertilizeoutput
+
+examplesummarizefertilizationoutput<-summarizefertilizationoutput
+
+examplesummarizepotentialpupoutput<-summarizepotentialpupoutput
+
+exampledesiredvec<-desiredvec
+
+examplecanwegetalltheallelesfromonecross<-canwegetalltheallelesfromonecross
+
+examplepointsperpupoutput<-pointsperpupoutput
+
+examplewhichpairstobreed<-whichpairstobreed
+#######
+
+#
+# usethis::use_data(exampleexampleData)
+# usethis::use_data(examplemeiosisoutput)
+#
+# usethis::use_data(examplecompilegametesoutput)
+#
+# usethis::use_data(examplesperm)
+#
+# usethis::use_data(exampleeggs)
+#
+# usethis::use_data(examplefertilizeoutput)
+#
+# usethis::use_data(examplesummarizefertilizationoutput)
+#
+# usethis::use_data(examplesummarizepotentialpupoutput)
+#
+# usethis::use_data(exampledesiredvec)
+#
+# usethis::use_data(examplecanwegetalltheallelesfromonecross)
+#
+# usethis::use_data(examplepointsperpupoutput)
+#
+# usethis::use_data(examplewhichpairstobreed)
 ### FIN ###
