+++ man/whichpairsshouldibreed.Rd
@@ -18,5 +18,5 @@ the crosses that produce the highest probability of desired pup or pups that are
 This function assists the user in deciding which breeding scheme will produce the result most likely to give them the mouse of their preferred genotype
 }
 \examples{
-whichpairsshouldibreed(x=pointsperpupoutput,desiredvector=desiredvec)
+whichpairsshouldibreed(x=examplepointsperpupoutput,desiredvector=exampledesiredvec)
 }
