### 2024-03-04 #6214 @csoneson
👋🏻 @RobLBaker - would you be interested in reviewing this submission for JOSS?

> MousebreedeR: A novel software to assist in the design of breeding schema for complex genotypes of experimental organisms

https://github.com/openjournals/joss-reviews/issues/6214

Thanks in advance!
