### 2024-03-05 11eccbd
fixed typo via roxygenise in example
