### 2024-03-05 3b65502
fixed typo in example
