### 2024-03-05 #6214 @sportiellomike
@yjunechoe, thank you for your first round of suggestions. I believe I have responded to the ones you requested (1 and 3). Thanks to your advice, I have removed the dependency of tidyverse, created example data that should help users when they look up functions, cleaned up a couple of miscellaneous things, and retested.
