### 2024-03-08 #6214 @RobLBaker
Unfortunately I also don't have the time at the moment. In addition, github is being blocked by my work computer.

-Rob

Robert Baker, Ph.D. (he/him)
Data Scientist/Integrative Biologist
National Park Service Inventory & Monitoring Division
1201 Oakridge Drive, Suite 150
Fort Collins, CO 80525
970-267-7216  (office)
From: Tyler Chafin ***@***.***>
Sent: Friday, March 8, 2024 9:26 AM
Cc: Baker, Robert L ***@***.***>; Mention ***@***.***>
Subject: [EXTERNAL] Re: [openjournals/joss-reviews] [PRE REVIEW]: MousebreedeR: A novel software to assist in the design of breeding schema for complex genotypes of experimental organisms (Issue #6214)

 This email has been received from outside of DOI - Use caution before clicking on links, opening attachments, or responding.

Sorry, I can't take on another review just now. I suggest igorkf as a possible alternative

-
