### 2024-03-08 #6214 @csoneson
👋🏻 @igorkf - would you be willing to review this submission for the Journal of Open Source Software (JOSS)?

> MousebreedeR: A novel software to assist in the design of breeding schema for complex genotypes of experimental organisms

https://github.com/openjournals/joss-reviews/issues/6214

Our review takes places openly on GitHub and is checklist-based, see [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more details.

Many thanks in advance!
