### 2024-03-08 #6214 @greggj2016
Sorry, I can't do that right now.

On Fri, Mar 8, 2024 at 11:09 AM Charlotte Soneson ***@***.***>
wrote:

> 👋🏻 @greggj2016 <https://github.com/greggj2016>, @dwinter
> <https://github.com/dwinter>, @tkchafin <https://github.com/tkchafin> -
> would you be willing to review this submission for the Journal of Open
> Source Software (JOSS)?
>
> MousebreedeR: A novel software to assist in the design of breeding schema
> for complex genotypes of experimental organisms
>
> #6214 <https://github.com/openjournals/joss-reviews/issues/6214>
>
> Many thanks in advance!
>
