### 2024-03-26 #6474 @yjunechoe
Hi @csoneson - I'm returning to this review after my first pass at the checklist (I have been awaiting a response from the author about automated tests before I proceed further)

I took some time to read the contents of this package more closely, and I have some serious concerns about the **Substantial scholarly effort** criterion.

Inside the package I count 12 functions - a combined LOC of 273 lines and an average of 23 lines for each function (LOC range: 4-45). Looking at internal function calls as a proxy for complexity, each function in the package is composed of an average 13 calls to other functions, mostly `c()`, `return()`, and `print()`. See below for the code involved in this analysis.

I'm afraid there's not much for me to work with here to evaluate the package. As per the [review guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html), I would like to flag the submission as questionable scope before proceeding further.

---

<summary>Analysis of substantial scholarly effort</summary>

``` r
## Setup
functions <- c(
  "addmouseID",
  "addpoints",
  "canwegetalltheallelesfromonecross",
  "compilegametes",
  "engageinmeiosis",
  "fertilize",
  "pointsperpup",
  "possiblepointsfromdesiredoutcome",
  "spermandeggs",
  "summarizefertilization",
  "summarizepotentialpups",
  "whichpairsshouldibreed"
)
length(functions)
#> [1] 12

endpoint <- "https://raw.githubusercontent.com/sportiellomike/mousebreedeR/main/R/"
files <- file.path(endpoint, paste0(functions, ".R"))
lines <- lapply(files, readLines)

## Code lines
count_code_lines <- function(x) {
  code <- grep("^(#|\\s*$)", x, invert = TRUE, value = TRUE)
  length(code)
}
all_code_lines <- sapply(lines, count_code_lines)
sum(all_code_lines)
#> [1] 273
mean(all_code_lines)
#> [1] 22.75
range(all_code_lines)
#> [1]  4 45

## Function complexity
parsed <- lapply(lines, \(x) parse(text = paste(x, collapse = "\n")))
parse_data <- do.call(rbind, lapply(parsed, getParseData))
function_calls <- parse_data |>
  subset(token == "SYMBOL_FUNCTION_CALL" & !text %in% functions) |>
  el("text") |>
  table()
head(sort(function_calls, decreasing = TRUE))
#>
#>  which      c return  print subset length
#>     18     13     12      9      8      7
sum(function_calls)
#> [1] 152
sum(function_calls)/length(functions)
#> [1] 12.66667
```
