### 2024-03-26 #6474 @sportiellomike
- Apologies @yjunechoe, I wasn't aware you were waiting on me. I was waiting for your full review to come back before working on your recommendations.
- The package itself was created to solve a problem: help users sacrifice fewer mice, get the needed genotypes for given experiments in fewer crosses to accelerate research, and do so in a convenient way usable with any number of loci not constricted to one genetic model or organism. I believe we have accomplished that goal. The vignette walks through how it achieves this goal, in my opinion. I can't speak to whether solving those problems meets the threshold of the journal, but I do believe the package achieves the goals it set out to solve.
