### 2024-03-26 #6474 @yjunechoe
Hi @sportiellomike - thanks for the quick response, and I apologize for not being clear on the issue I opened on automated tests

To your points - I understand your perspective on the usefulness of the tool (albeit I've not fully evaluated this formally). I only mean to comment on the structure and contents of the package as an open source software developed for general consumption, according to the scope and standards of JOSS specifically. To clarify, I'm following the practice of previous reviewers who have raised issues of LOC and complexity as part of evaluating the substantial scholarly effort criterion (see [similar issues](https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue+is%3Aclosed+label%3AR+label%3Aquery-scope+%22substantial+scholarly+effort%22)).

I'm happy to complete the review and help with the review process up to a successful publication, but substantial scholarly effort is a rather tricky matter to resolve, especially if the package is already considered to be at a feature-complete state that addresses its goals. I only seek clarification from the editorial team - I apologize again for the interruption.
