### 2024-04-12 #6474 @jsun
This package contains simple but important features to assist wet researchers. Therefore, I think, the main target users of this package would be wet researchers who are unfamiliar with coding programmes. Therefore, I recommend the author to create easy-to-use functions and clear documentation (vignette). The following comments are based on a check of this contribution.

- package
  - The function names in this package are long and difficult to read. If possible, please rename the function names (e.g.,`summarizepotentialpups` to `summarize_potential_pups` or `summarizePotentialPups`). Alternatively, please consider using the S4 class to develop this package. In this case, instead of using `summarizepotentialpups`, a simple `summary` function can be defined.
  - If visualisation is one of the core features of this package, implement a `plot` function to reduce end-user costs.
  - Please consider changing the output format of `canwegetalltheallelesfromonecross` and `whichpairsshouldibreed` functions. These outputs are not R-like.

- vignette
  - Many researchers who are not familiar with R do not know how to obtain vignettes, so please indicate how to obtain vignettes in README.md (perhaps after the 'How to install' section).
  - Please add an 'Abstract' to the vignette to explain why this package is needed (which may overlap with the 'state of need' in the manuscript) and what this vignette describes.
  - The figures in the vignette are too small to check. Please adjust the size of the figure (e.g., set the width to 100%).
  - The vignette contains `plot<-ggplot(subsetmelt)+geom_col(aes(y=percent,x=genotype,fill=genotype))+facet_wrap(~gene)+` kind of code. The `plot` is the famous generic function, please consider using other variable name to assign the ggplot object.
  - Coding rules are not uniform. For example, `whichpairstobreed<-whichpairsshouldibreed(x=pointsperpupoutput,desiredvector = desiredvec)`; there is no spaces between `=` and other contains the spaces.
  - Some variable names are too long to catch the information. Please consider use shorter one or undanstandable one (e.g., `compilegametesoutput` to `compilegametes_output`)
