### 2024-04-15 #6474 @csoneson
> @csoneson: the reviewer @yjunechoe encouraged me to comment here regarding the "substantial scholarly effort" guideline in the other thread.

Thanks @sportiellomike - the editorial team is discussing the submission and we should have an answer for you soon! Thanks for your patience.
