### 2024-04-16 #6474 @yjunechoe
I expect some time with back-and-forths on the package repository to get the submission into a good shape, and I'll keep raising specific points as GH issues, but in the meantime here are my general thoughts on the package:

First off I agree with all points that the other reviewer has mentioned above. I'd especially like to echo the point about function/variable names being difficult to parse, the general inconsistencies in coding style (use of spaces and indentation, use of semicolon, etc.), and the fact that the return value of some functions are not R-like. My one point of digression from the other reviewer is that for a package of this size, I don't think a S4 class is necessary, but I do agree that some kind of a custom print method is appropriate (perhaps the lighter-weight S3).

And two big-picture comments/questions of my own:

## 1) Non-idiomatic R code

There are several cases of non-idiomatic R code in the package. The other reviewer has already given an example of one: the `canwegetalltheallelesfromonecross()` function returns a string that's either `"notonecrossforonecopy"` or `"onecrossforonecopy"`, which is very difficult to program over. It'd be more appropriate to have the function be called something like `is_onecross()` that returns `TRUE` or `FALSE`.

A few other examples to this larger point about non-idiomatic R code:

Inside `fertilize()`, I see:

```r
  fertilizeoutput<<-listofzygotes
```

With no local binding for the variable `fertilizeoutput`. This inevitably assigns into the user environment - such leaky behaviors (and side-effect functions in general) are not R-like and highly discouraged for a function in a package. This also happens in `spermandeggs()`

Also inside `fertilize()` and elsewhere, I see:

```r
which(spermcolnumerics == T)
```

This should simply be `which(spermcolnumerics)` - equality checks to a scalar boolean are redundant and can introduce silent type coercion bugs, since `==` is loose equality and R doesn't have vectorized strict equality. This comment also applies to use of `isTRUE()` in a few `if` statements elsewhere.

Additionally, in `addpoints()` and elsewhere, I see:

```r
  pointsvector<-c()
  for (k in 1:length(desiredvector)) {
```

But growing a vector is costly in R. The appropriate pattern is to either pre-allocate the vector and fill it with values, or `*apply()` over the iterator.

Lastly to keep this short, in `pointsperpup()` and elsewhere, I see:

```r
  genecolx[genecolx=='homopos']<-2
  genecolx[genecolx=='het']<-1
  genecolx[genecolx=='homoneg']<-0
```

Such incremental replacement of values are odd when the type doesn't match. The idiomatic R approach would be vectorized subsetting of a named vector, like:

```r
genotypes <- setNames(0:2, c("homoneg", "het", "homopos"))
genotypes[genecolx]
```

I'd also like to note that the above code chunk of applying number coding to genotypes is copy-pasted across several function calls. To this point, I'd like to also point out that the package has zero internal utility functions (all functions are exported), which is uncharacteristic of a package.

These are perhaps minor points and there's quite some way to go on this front, but I would like to see these issues resolved for a publication-quality software

## 2) Who is the target audience for this tool?

I'm a little puzzled at the level of R experience that is expected of the average user of this tool. Currently, I sense bit of a tension in what is assumed of the audience: the audience is assumed to be familiar enough with R to be able to build/parse the vignettes and understand the package from reading and running the examples themselves, but not familiar enough to such that we have functions like this that simply multiplies the length of a vector by 100, hidden under a function.

```r
possiblepointsfromdesiredoutcome<-function(desiredvector = desiredvec){
  possiblepoints<-length(desiredvector)*100
  return(possiblepoints)
}
```

Overall, there is not enough hand-holding for a novice R user, but enough non-idiomatic R quirks to confuse a more experienced R user. It would be beneficial for design of the package (and also for the paper) to clearly state the level of R knowledge assumed/required of the user of this package, and cater specifically to that audience (and broaden the audience once the package meets the need of its core users).

I get the sense that it's the former (package is for novice programmers), and I actually wonder in that case if something more like a point-and-click interface is more appropriate (like a shiny app complement to the software). I encourage the author to at least look into this somewhat more obvious method of supporting novice users and comment on the choice to stick to coding in R as the primary interface to this tool.
