### 2024-04-25 #6474 @sportiellomike
I am working on these issues and will tag you when resolved. As you mentioned, this will take some time. I appreciate your patience. I'd like to thank the reviewers for their thorough comments, which I know will make this software better and more usable.

Thank you again.
