+++ R/whichpairsshouldibreed.R
@@ -1,16 +1,16 @@
-#' whichpairsshouldibreed
+#' which_pairs_should_i_breed
 #'
 #' @description This function assists the user in deciding which breeding scheme will produce the result most likely to give them the mouse of their preferred genotype
-#' @param x the output of the pointsperpup function
+#' @param x the output of the points_per_pup function
 #' @param desiredvector a character vector of the genotype you want to get to
 #'
 #' @return the crosses that produce the highest probability of desired pup or pups that are closest to the desired genotype.
 #' @export
 #'
 #' @examples
-#' whichpairsshouldibreed(x=examplepointsperpupoutput,desiredvector=exampledesiredvec)
+#' which_pairs_should_i_breed(x=examplepointsperpupoutput,desiredvector=exampledesiredvec)
 
-whichpairsshouldibreed<-function(x=pointsperpupoutput,desiredvector=desiredvec){
+which_pairs_should_i_breed<-function(x=pointsperpupoutput,desiredvector=desiredvec){
   maxpoints<-(max(x$points))
   maxpointindex<-which(x$points==maxpoints)
   maxpointsub<-subset(x,x$points==maxpoints)
@@ -25,7 +25,7 @@ whichpairsshouldibreed<-function(x=pointsperpupoutput,desiredvector=desiredvec){
   print(crosseswithmaxpointsmaxpercent)
   print('which gives you the following percent chance of making that pup:')
   print(maxpercent)
-  if (canwegetalltheallelesfromonecross(x = fertilizeoutput,desiredvector)=='notonecrossforonecopy') {
+  if (can_we_get_all_the_alleles_from_one_cross(x = fertilizeoutput,desiredvector)=='notonecrossforonecopy') {
     print('You cannot get all the alleles you want into one mouse with one cross (even as a heterozygote), so you should make crosses from the list above that get as many alleles into the same pup as possible, and eventually cross those pups together.')
   }
   return(crosseswithmaxpointsmaxpercent)
