+++ R/pointsperpup.R
@@ -1,4 +1,4 @@
-#' pointsperpup
+#' points_per_pup
 #'
 #' @description Calculates how many points a potential pup would be worth according to its genotype.
 #' @param x summarizepotentialpupoutput
@@ -8,9 +8,9 @@
 #' @export
 #'
 #' @examples
-#' pointsperpup(x=examplesummarizepotentialpupoutput,desiredvector=exampledesiredvec)
+#' points_per_pup(x=examplesummarizepotentialpupoutput,desiredvector=exampledesiredvec)
 
-pointsperpup<-function(x,desiredvector=desiredvec){
+points_per_pup<-function(x,desiredvector=desiredvec){
   genecols <- c()
   for (t in colnames(x)) {
     if (isTRUE(any(x[t] == genotypevalschar))) {
@@ -22,8 +22,8 @@ pointsperpup<-function(x,desiredvector=desiredvec){
   genecolx[genecolx=='het']<-1
   genecolx[genecolx=='homoneg']<-0
   genecolx <- as.data.frame(lapply(genecolx,as.numeric))
-  genecolx$possiblepoints<- apply(X=genecolx,MARGIN = 1,possiblepointsfromdesiredoutcome)
-  genecolx$points <- apply(X=genecolx,MARGIN = 1,addpoints)
+  genecolx$possiblepoints<- apply(X=genecolx,MARGIN = 1,possible_points_from_desired_outcome)
+  genecolx$points <- apply(X=genecolx,MARGIN = 1,add_points)
   genecolx$normalizedpoints<- 100*(genecolx$points / genecolx$possiblepoints)
   y<-cbind(genecolx,x[,!names(x)%in%genecols])
   return(y)
