+++ R/addmouseID.R
@@ -1,4 +1,4 @@
-#' addmouseID
+#' add_mouse_ID
 #'
 #' @description Adds mouseID to the original dataframe provided
 #' @param x dataframe to add column to.
@@ -7,10 +7,10 @@
 #' @export
 #'
 #' @examples
-#' addmouseID(exampleexampleData)
+#' add_mouse_ID(exampleexampleData)
 
 
-addmouseID <- function(x) {
+add_mouse_ID <- function(x) {
   x$mouseID <- paste0('mouse', 1:dim(x)[1])
   return(x)
 }
