+++ man/compilegametes.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/compilegametes.R
-\name{compilegametes}
-\alias{compilegametes}
-\title{compilegametes}
-\usage{
-compilegametes(x)
-}
-\arguments{
-\item{x}{the output of engageinmeiosis()}
-}
-\value{
-a dataframe of gametes
-}
-\description{
-This finishes the create of the gametes.
-}
-\examples{
-compilegametes(examplemeiosisoutput)
-}
