+++ man/summarize_fertilization.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/summarizefertilization.R
-\name{summarizefertilization}
-\alias{summarizefertilization}
-\title{summarizefertilization}
+% Please edit documentation in R/summarize_fertilization.R
+\name{summarize_fertilization}
+\alias{summarize_fertilization}
+\title{summarize_fertilization}
 \usage{
-summarizefertilization(x)
+summarize_fertilization(x)
 }
 \arguments{
 \item{x}{the output of fertilize()}
@@ -16,5 +16,5 @@ Dataframe with columns to summarize the probabalistic outcome of fertilizations,
 summarizes the overall picture of fertilization.
 }
 \examples{
-summarizefertilization(examplefertilizeoutput)
+summarize_fertilization(examplefertilizeoutput)
 }
