### 2024-05-19 39486ed
some updates to vignette, added invisible to can we get all alleles from one cross
