### 2024-05-19 e8b1b50
Addressing reviewer comments below
Changed which(*** =T) to which(***)

Removed <<- operator

Changed vignette plot width/height
