### 2024-05-19 7eab457
Update engage_in_meiosis.R
reverting to which(***==T) after vignettes won't run
