### 2024-05-19 5f44f73
Update fertilize.R
reverting to the which ==T again cuz vignette broke (see last commit)
