### 2024-05-20 fbd2fd9
changed spacing, updated readme, updated vignette with plot_to_plot and abstract per reviewers
