+++ R/add_points.R
@@ -11,26 +11,26 @@
 #' add_points(x,desiredvector=exampledesiredvec)
 
 add_points<-function(x,desiredvector=desiredvec){
-  desiredvector[desiredvector=='homopos']<-2
-  desiredvector[desiredvector=='het']<-1
-  desiredvector[desiredvector=='homoneg']<-0
+  desiredvector[desiredvector == 'homopos']<-2
+  desiredvector[desiredvector == 'het']<-1
+  desiredvector[desiredvector == 'homoneg']<-0
   desiredvector<-as.numeric(desiredvector)
   length(desiredvector)
   pointsvector<-c()
   for (k in 1:length(desiredvector)) {
-    if (x[k]==desiredvector[k]) {
+    if (x[k] == desiredvector[k]) {
       pointstoaddtovector<-100
       pointsvector <- c(pointsvector, pointstoaddtovector)
     }
-    if (x[k]-desiredvector[k]==-1) {
+    if (x[k]-desiredvector[k] == -1) {
       pointstoaddtovector<-50
       pointsvector <- c(pointsvector, pointstoaddtovector)
     }
-    if (x[k]-desiredvector[k]==1) {
+    if (x[k]-desiredvector[k] == 1) {
       pointstoaddtovector<-50
       pointsvector <- c(pointsvector, pointstoaddtovector)
     }
-    if (abs(x[k]-desiredvector[k])==2) {
+    if (abs(x[k]-desiredvector[k]) == 2) {
       pointstoaddtovector<-0
       pointsvector <- c(pointsvector, pointstoaddtovector)
 
