+++ vignettes/Vignette.Rmd
@@ -8,6 +8,13 @@ vignette: >
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
+
+# Abstract
+
+Efficient breeding is a scientific imperative: with a limited period of fertility in experimental organisms, users may only have limited attempts with organisms of desired genotypes [@RN204] [@RN203] since the time of fertility is often the limiting factor in experimental organisms. An inefficient cross can set the user back months in the case of a female pregnant with the non-optimal male's offspring once both gestational time and the necessary time it takes to wean pups off their mother's milk. Furthermore, as most research centers pay per diem costs, inefficiency is a large direct cost for many experimentalists, with individual cages often costing hundreds of dollars per year just in maintenance fees, not even accounting for the cost of extra rounds of genotyping for which one must account for genotyping costs (PCR primers, PCR mastermix, gels, dye, etc) or for people's time, which may take the form of decreased data production and the cost of of actually paying employees to do this work. Ethical and animal welfare concerns also arise as inefficient breeding results in the mass euthanasia of non-desired animals as well as ear/tail-clipping required to genotype which may cause animals pain and distress.
+
+While the breeding schema to obtain a litter of full knockout mice from one Wildtype/Wildtype parent and one knockout/knockout parent is straightforward, this is not the case for when one has 4 alleles at 4 loci in 4 separate mice that need to be in the same mouse for a given planned experiment, for example. Furthermore, no current software exists to our knowledge that can quantitatively assist the user in creating their breeding schema. Indeed, if there are 2 alleles at each locus, when attempting to make a genetically marked, inducible, cre-lox model with T cell specificity as our lab was doing, prompting us to create this software, 81 possible combinations exist (3 (AA, Aa, and aa) to the 4th power). With 81 possible males and 81 possible females to mate, 6561 possible pairings exist (81 * 81). Use of this mousebreedeR software to optimize that breeding schema resulted in the publication of a thesis project and manuscript being prepped for submission. 
+
 # Walkthrough
 First we start off by loading all of the necessary libraries.
 ```{r message=FALSE, warning=FALSE}
@@ -152,7 +159,7 @@ This chunk of code will plot separate plots for each gene, and show you the dist
 for (w in uniquemeltgenes) {
   subsetmelt<-subset(meltsummarizefertoutput,meltsummarizefertoutput$gene==w)
   print(subsetmelt)
-  plot<-ggplot(subsetmelt)+geom_col(aes(y=percent,x=genotype,fill=genotype))+
+  plot_to_plot<-ggplot(subsetmelt)+geom_col(aes(y=percent,x=genotype,fill=genotype))+
     facet_wrap(~momdad)+
     scale_fill_viridis(labels=c("freqhomoneg" = "HomozygousNeg", "freqhet" = "Heterozygous",
                                 "freqhomopos" = "HomozygousPos"),discrete = T,end=.8)+
@@ -160,7 +167,7 @@ for (w in uniquemeltgenes) {
     theme(axis.text.x = element_text(angle=90),
           plot.title = element_text(hjust = 0.5))
 
-  plotlist[[length(plotlist) + 1]] <- plot
+  plotlist[[length(plotlist) + 1]] <- plot_to_plot
 
 }
 ```
@@ -179,7 +186,7 @@ uniquemepairings<-unique(meltsummarizefertoutput$momdad) # pull the unique pairi
[…]

[…]
