+++ README.md
@@ -36,8 +36,8 @@ This software is built for people of intermediate R experience. Having said that
 * You can test out the supplied example data by clicking "Run" and having done everything described in the "How to install" and "How to actually run the software" instructions on the ReadMe document you're actively rading right now.
 * To use the program on your own data, supply your own dataframe in the correct format. 
 	* The format should be in the same format as the CSV file supplied to you in this github called 'edit_this_CSV_with_breeder_mice.csv'. You can download this file at: https://github.com/sportiellomike/mousebreedeR/blob/JOSS_Edits/edit_this_CSV_with_breeder_mice.csv. 
-	* You can directly edit this file and then re-save it named as something else. Then, in the vignette on line 45, just replace the file that's in quotes with the path of your file.
-* Make sure to tell the program what your desired genotype is. You can do this on line 115. The length of this vector should be the same number of gene columns in your breeder column you described. Said another way, if you need 3 alleles to be homozygous positive, the dataframe of breeders should be 3 gene columns (and a sex column) and line 115 should be: `desiredvec<-c('homopos', 'homopos', 'homopos')`
+	* You can directly edit this file and then re-save it named as something else. Then, in the vignette on line 52, just replace the file that's in quotes with the path of your file.
+* Make sure to tell the program what your desired genotype is. You can do this on line 122. The length of this vector should be the same number of gene columns in your breeder column you described. Said another way, if you need 3 alleles to be homozygous positive, the dataframe of breeders should be 3 gene columns (and a sex column) and line 122 should be: `desiredvec<-c('homopos', 'homopos', 'homopos')`
 * Once you're all set, click the "Run" button on the top middle of the screen (top right of top left pane). 
 * Alternatively, you can click "Knit" near the middle of the top left pane and scroll through your output there. It is a little bit easier to read.
 
