### 2024-05-20 f6050b1
contributor and code of conduct added
