### 2024-05-20 516e73e
Update add_points.R
moved previous example down since there is no reason for it to have example
