### 2024-05-20 97bc28d
added check yaml, trying to fix example for add_points
