### 2024-05-20 #6474 @sportiellomike
Thank you to all involved for all your help. I'll get to it.

- @jsun's comments
- I have changed all the names as requested. I agree this is much better now (more readable, more sensible, more in line with standards). Thank you.
- Visualization I actually removed before submission, and only added it back in in case reviewers asked for it. I really don't think it adds much to the project.
- I changed the output formats as requested and have altered the functions slightly so they only output desired text. While they still out put text to the user, I can't think of another way that the user could get this information. To me, the best output is direct text informing user of the advice of the software, which is the primary function of the software.
- Totally agree that I needed to explain how to obtain vignettes. This has been completed.
- I have added abstract as well
- I have changed the size of the figures.
- I have changed the "plot" to a different variable name as requested
- I have fixed spacing
- I have changed readability of output variables as requested
- Thank you so much for all your help

- @yjunechoe
- I have corrected the `canwegetalltheallelesfromonecross()` function using the invisible function appropriately as suggested in by other users that had similar situations.
- I have removed the unnecessary <<- throughout the package. However, given the linear nature of this software, I elected to keep the sperm and egg one instead of assigning the two outputs for male and female to user supplied variables as I think there is a high likelihood of confusion by the user of which is egg and which is sperm. In order to make that foolproof, I kept the <<- (but removed the other case of it as I totally agree with you it was non-optimal)
-  With regards to `which(spermcolnumerics == T)`, when I correct it to `which(spermcolnumerics)`, the whole thing breaks and the vignette can no longer build. I have changed it back to the original as a result.
- I'm confused why my assigning of 0, 1, and 2 is non-idiomatic. The function works appropriately and I think is more understandable to the user of what I'm doing (assigning the number of alleles directly over the text based ("homoneg", "het", "homopos").
- The function is designed for a step above novice, let's say intermediate. I have attempted to address some of the difficulties of user experience in the readme and vignette as suggested by both of you reviewers.

Thank you again for both reviewers and the editing team for your knowledge, advice, and assistance through this process.
