### 2024-05-20 #6474 @yjunechoe
@sportiellomike Thank you for the updates, here and elsewhere, and the hard work that went into them. I'm satisfied with your edits and/or justifications to the points I raised in my previous comment. Should they raise a separate problem down the line I will raise them again, but we can move forward on the other points (the ones I've opened as issues on the package repo).

I'm also happy to hear the addition of a shiny app - this will make the submission much stronger. The app should live inside `/inst/<APP NAME>/`, and a function should be added/exported to allow users to run the app through a function call. Please feel free to ping me on your repo in a new issue for the shiny app, once you have a minimally working app and have put it under `/inst`
