+++ R/which_pairs_should_i_breed.R
@@ -21,12 +21,13 @@ which_pairs_should_i_breed<-function(x = pointsperpupoutput,desiredvector = desi
   crosseswithmaxpointsmaxpercent<-unique(maxpointsub$momdad[maxpercentindex])
   print('These following cross(es) get you closest to your desired mouse, and may even give you your desired mouse:')
   print(crosseswithmaxpoints)
-  print('The cross(es) that gives you the highest percentage of pups with that closest genotype is:')
+  print('See the table below for probabilities of desired pups.')
+  print('The cross(es) that gives you the highest percentage of pups with genotypes most ideal for breeding if your ideal mouse cannot be made in one cross (preferring homopos over het) is:')
   print(crosseswithmaxpointsmaxpercent)
-  print('which gives you the following percent chance of making that pup:')
+  print('which gives you the following percent chance of making that pup ideal for breeding (preferring homopos over het):')
   print(maxpercent)
   if (can_we_get_all_the_alleles_from_one_cross(x = fertilize_output,desiredvector) == 'notonecrossforonecopy') {
     print('You cannot get all the alleles you want into one mouse with one cross (even as a heterozygote), so you should make crosses from the list above that get as many alleles into the same pup as possible, and eventually cross those pups together.')
   }
-  return(crosseswithmaxpointsmaxpercent)
+  # return(crosseswithmaxpointsmaxpercent)
 }
