### 2024-05-21 #6474 @sportiellomike
@yjunechoe shiny app is live. if you choose to try it, I put another example datasets up there to test it out that are less complex than original (edit_this_CSV_with_breeder_mice2.csv).

I included "visualization" of uploaded csv for user to check to make sure it looks right, "visualization" of desired genotype that the user can manually type in to make sure it looks correct, the advice the program gives to the user, a couple relevant tables for potential pups/summary per gene and cross, and a download button for the user to download potential pup table.

Let me know if I misunderstood your instructions on where to put it.
