### 2024-05-22 #6474 @jsun
@sportiellomike thank you for considering my comments. Here is my other suggestion (just suggestions), you don't need to change your package as in this suggestion.

In the case of the `can_we_get_all_the_alleles_from_one_cross` function, the function name is "can we get ... ", like a question , so it is just Yes or No as an answer. Therefore, I think `TRUE` or `FALSE` would be appropriate as the return value of this function. Or, change the function name to `get_all_alleles`. This function can return the number of crossing experiments needed to get all alleles as an integer.

For `which_pairs_should_i_breed` function, the return may organized as a list or slots, for example:

```
which_pairs_should_i_breed <- function() {

...

  output <- list(recommend_1 = NULL, recommend_2 = NULL, prob=NA, all_alleles_with_one_corss=NA)

  output$recommend_1 <- data.frame(
    mom = c('mouse1', 'mouse1', 'mouse1', 'mouse1', 'mouse2', ...),
    dad = c('mouse5', 'mouse6', 'mouse7', 'mouse8', 'mouse5', ...)
  )
  output$recommend_2 <- data.frame(
    mom = c('mouse1', 'mouse1', 'mouse1', 'mouse1', 'mouse2', ...),
    dad = c('mouse5', 'mouse6', 'mouse7', 'mouse8', 'mouse5', ...)
  )
  output$prob <- 0.25
  output$all_alleles_with_one_corss <- FALSE

  output
}
```

For some outputs, I can only see `momdad` column in the data.frame. If it possible, use `mom` and `dad` columns instead of using `momdad` column (of course, you can still remain the `momdad` column). Adding `mom` and `dad` columns would facilitate secondary use of the outputs.
