+++ R/shiny_mousebreedeR.R
@@ -1,0 +1,260 @@
+#' fertilize
+#'
+#' @description Use the interactive shiny app to run the mousebreedeR software.
+#'
+#' @examples
+#' shiny_mousebreedeR()
+#'
+shiny_mousebreedeR <- function() {
+
+  library(shiny)
+  library(shinythemes)
+  library(mousebreedeR)
+  library(dqrng)
+  library(dplyr)
+  library(gtools)
+  library(ggplot2)
+  library(reshape2)
+  library(viridis)
+  library(ggpubr)
+  `%!in%` <- Negate(`%in%`)
+
+  # Define UI for application that draws a histogram
+  ui <- fluidPage(theme = shinytheme("slate"),
+                  # Application title
+                  titlePanel("mousebreedeR"),
+
+                  # Sidebar with a slider input for number of bins
+                  sidebarLayout(
+                    sidebarPanel(
+                      h4('Further documentation for this interactive can be found at https://github.com/sportiellomike/mousebreedeR'),
+                      h2('Breeder mice upload'),
+                      fileInput("file1", "Choose CSV File", accept = ".csv"),
+                      h2('Desired genotype'),
+                      textInput('desiredvec_shinyinput',"Input your desired vec here with commas, like this using only some combination of homopos,het, or homoneg: homopos,het,homoneg"),
+                      h2('Download potential pup table'),
+                      downloadButton("downloadData", "Download potential pup table"),
+                      # selectInput("species", "species",
+                      #             c('Mouse'='Mmu',
+                      #               'Human'='Hsa')),
+                      # selectInput('geneformat','Gene format',
+                      #             c('Symbol','ENTREZID')),
+                      # selectInput('inputformat','Input format',
+                      #             c('Dataframe')),
+                      # selectInput("padjcolname", "Column to use for P value adjustment",''),
+                      # selectInput("LFCcolname", "Column with adjusted Log Fold Change values. Note that LFC col name must be log2FoldChange.",''),
+                      # numericInput("pcutoff", "Significance cutoff (alpha)", 0.05, min = 0, max = 1),
+                      # downloadButton("downloadData", "Download output table"),
+                      # h1('This interactive supports dataframe inputs to fluximplied only.')
+                    ),
+
+                    #create the main panel plotting and printing outputs
+                    mainPanel(
+                      # h2("Output table"),
+                      # tableOutput(outputId = 'table'),
+                      # h2("Plot"),
+                      # plotOutput(outputId = 'plot'),
+
+                      h2("Breeder Mice"),
+                      h4("This is the CSV you supplied to us. Make sure the column on the right is 'sex' and the columns to the left are your gene columns. There should not be a separate column for row names in your CSV."),
+                      h4(dataTableOutput(outputId = 'suppliedCSV')),
+
+                      h2("Desiredvec"),
+                      h4('This is the desiredvec you entered to the left. Make sure the only values here are homopos, homoneg, or het separated by commas. Each gene in your supplied csv should have a desired geneotype at each locus (for example, 3 gene columns should have 3 entries to the left separated by commas).'),
+                      verbatimTextOutput("desiredvec_processed"),
+
+                      h2("Number of crosses required"),
[…]
