### 2024-05-23 45f224a
Update points_per_pup.R
