### 2024-05-23 a410570
points per pup is causing error in shiny
