### 2024-05-23 c4d5adf
Revert "points per pup is causing error in shiny"
This reverts commit a4105704cec557376c64273b07dbee6a1990eac4.
