### 2024-05-23 8ab13d1
Update app.R
FIXED THE BUG in shiny
