### 2024-05-23 6631c95
new stuff
finished shiny app, made shiny function, made documentation for it, made new example dataset
