+++ R/shiny_mousebreedeR.R
@@ -1,11 +1,12 @@
-#
-# This is a Shiny web application. You can run the application by clicking
-# the 'Run App' button above.
-#
-# Find out more about building applications with Shiny here:
-#
-#    https://shiny.posit.co/
-#
+#' fertilize
+#'
+#' @description Use the interactive shiny app to run the mousebreedeR software.
+#'
+#' @examples
+#' shiny_mousebreedeR()
+#'
+shiny_mousebreedeR <- function() {
+
 
 library(shiny)
 library(shinythemes)
@@ -288,6 +289,4 @@ server <- function(input, output) {
 
 
 }
-
-# Run the application
-shinyApp(ui = ui, server = server)
+}
