+++ man/shiny_mousebreedeR.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/shiny_mousebreedeR.R
+\name{shiny_mousebreedeR}
+\alias{shiny_mousebreedeR}
+\title{fertilize}
+\usage{
+shiny_mousebreedeR()
+}
+\description{
+Use the interactive shiny app to run the mousebreedeR software.
+}
+\examples{
+shiny_mousebreedeR()
+
+}
