+++ R/add_mouse_ID.R
@@ -7,7 +7,19 @@
 #' @export
 #'
 #' @examples
-#' add_mouse_ID(exampleexampleData)
+#' library(shiny)
+#' library(shinythemes)
+#' library(mousebreedeR)
+#' library(dqrng)
+#' library(dplyr)
+#' library(gtools)
+#' library(ggplot2)
+#' library(reshape2)
+#' library(viridis)
+#' library(ggpubr)
+#' 
+#' head(exampleexampleData) # let's take a look at some example data before we run add_mouse_ID()
+#' add_mouse_ID(exampleexampleData) # check it out! We have a new column!
 
 
 add_mouse_ID <- function(x) {
