+++ R/notin.R
@@ -1,4 +1,4 @@
-#' %!in%
+#' %notin%
 #'
 #' @description Negates %in%
 #'
@@ -6,6 +6,19 @@
 #' @export
 #'
 #' @examples
-#' people %!in% distress
+#' library(shiny)
+#' library(shinythemes)
+#' library(mousebreedeR)
+#' library(dqrng)
+#' library(dplyr)
+#' library(gtools)
+#' library(ggplot2)
+#' library(reshape2)
+#' library(viridis)
+#' library(ggpubr)
 
-`%!in%` <- Negate(`%in%`)
+
+#' c(3,4,5) %in% c(4,5,6)
+#' c(3,4,5) %notin% c(4,5,6)
+
+`%notin%` <- Negate(`%in%`)
