+++ man/grapes-notin-grapes.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/notin.R
+\name{\%notin\%}
+\alias{\%notin\%}
+\title{%notin%}
+\usage{
+... \%notin\% NA
+}
+\value{
+negated %in%
+}
+\description{
+Negates %in%
+}
+\examples{
+library(shiny)
+library(shinythemes)
+library(mousebreedeR)
+library(dqrng)
+library(dplyr)
+library(gtools)
+library(ggplot2)
+library(reshape2)
+library(viridis)
+library(ggpubr)
+c(3,4,5) \%in\% c(4,5,6)
+c(3,4,5) \%notin\% c(4,5,6)
+}
