+++ man/can_we_get_all_the_alleles_from_one_cross.Rd
@@ -18,5 +18,21 @@ If a dataframe was supplied, it should also return a dataframe as well as a bar 
 This function will tell you if multiple crosses will be required to get the desired mouse.
 }
 \examples{
-can_we_get_all_the_alleles_from_one_cross(x=examplefertilizeoutput,desiredvector=exampledesiredvec)
+library(shiny)
+library(shinythemes)
+library(mousebreedeR)
+library(dqrng)
+library(dplyr)
+library(gtools)
+library(ggplot2)
+library(reshape2)
+library(viridis)
+library(ggpubr)
+head(exampleexampleData) # let's take a look at our example data
+engage_in_meiosis_output_example<-engage_in_meiosis(exampleexampleData) # Check it out! We can look the gene makeup of eggs and sperm now
+compile_gametes_output_example<-compile_gametes(engage_in_meiosis_output_example) # Meiosis step completed. Here are all the possible gametes from our breeder mice.
+sperm_and_eggs_output_example<-sperm_and_eggs(x=compile_gametes_output_example,sex='sex') # Saves the outputs of which gametes are sperm, and which are eggs.
+fertilize_output_example<-fertilize(malegametes = sperm,femalegametes = eggs) # create all potential pups from all possible pairings.
+desiredvec<-c('het','het','het','het') # the genotype of your desired mouse
+can_we_get_all_the_alleles_from_one_cross(x=fertilize_output_example,desiredvector=desiredvec)
 }
