### 2024-05-24 30f7b64
fixed libraries in shiny and example in add_points
