+++ man/which_pairs_should_i_breed.Rd
@@ -29,12 +29,18 @@ library(reshape2)
 library(viridis)
 library(ggpubr)
 head(exampleexampleData) # let's take a look at our example data
-engage_in_meiosis_output_example<-engage_in_meiosis(exampleexampleData) # Check it out! We can look the gene makeup of eggs and sperm now
-compile_gametes_output_example<-compile_gametes(engage_in_meiosis_output_example) # Meiosis step completed. Here are all the possible gametes from our breeder mice.
-sperm_and_eggs_output_example<-sperm_and_eggs(x=compile_gametes_output_example,sex='sex') # Saves the outputs of which gametes are sperm, and which are eggs.
-fertilize_output_example<-fertilize(malegametes = sperm,femalegametes = eggs) # create all potential pups from all possible pairings.
+# Check it out! We can look the gene makeup of eggs and sperm now
+engage_in_meiosis_output_example<-engage_in_meiosis(exampleexampleData) 
+# Meiosis step completed. Here are all the possible gametes from our breeder mice.
+compile_gametes_output_example<-compile_gametes(engage_in_meiosis_output_example) 
+# Saves the outputs of which gametes are sperm, and which are eggs.
+sperm_and_eggs_output_example<-sperm_and_eggs(x=compile_gametes_output_example,sex='sex') 
+# create all potential pups from all possible pairings.
+fertilize_output_example<-fertilize(malegametes = sperm,femalegametes = eggs) 
+fertilize_output<-fertilize_output_example
 desiredvec<-c('het','het','het','het') # the genotype of your desired mouse
-summarize_potential_pups_output<-summarize_potential_pups(fertilize_output_example) # take a look at the distributions of potential pups
+# take a look at the distributions of potential pups
+summarize_potential_pups_output<-summarize_potential_pups(fertilize_output_example) 
 points_per_pup_output_example<-points_per_pup(x=summarize_potential_pups_output,desiredvector=desiredvec)
 which_pairs_should_i_breed(x=points_per_pup_output_example,desiredvector=desiredvec)
 }
