### 2024-05-24 #6474 @sportiellomike
@jsun I fixed the mom and dad columns in the shiny as you recommended. Thanks!

I elected to defer your suggestion for the output lists not because I think it wouldn't work, just because several other functions depend on the output and would involve changing a long list of functions in the package. I will take the themes of your suggestion into account if I develop a version 2.0 and during future package development.
