### 2024-05-24 #6474 @sportiellomike
@jsun @yjunechoe

I believe I have addressed all of your necessary proposed changes, considered and changed package in light of most of your optional suggestions, and built a shiny for use as well. Please let me know if there is anything else I can do for you. Thank you again for reviewing this software and thank you @csoneson for editing.
