+++ README.md
@@ -27,7 +27,9 @@ To install, go to Rstudio or the R command line and run the following code.
 `library(mousebreedeR)`
 
 # How to actually run the software
-This software is built for people of intermediate R experience. Having said that, so long as you follow these instructions very carefully below, anyone should be able to use this software. Generally, I'd recommend just using the GUI unless you have a good reason not to (https://sportiellomike.shinyapps.io/mousebreedeR/[https://sportiellomike.shinyapps.io/mousebreedeR/])
+Generally, I'd recommend just using the GUI unless you have a good reason not to (https://sportiellomike.shinyapps.io/mousebreedeR/[https://sportiellomike.shinyapps.io/mousebreedeR/]) since it can be used by true novices, and is the quickest way to get results.
+
+This pacakge/vignette is built for people of intermediate R experience. Having said that, so long as you follow these instructions very carefully below, anyone should be able to use it. 
 
 ## Download what you need
 * Download the vignette from this link: https://github.com/sportiellomike/mousebreedeR/blob/main/vignettes/Vignette.Rmd
