### 2024-05-27 #6474 @yjunechoe
Thanks, I see the shiny app. Could you please also create a vignette of instructions for how to use the shiny app? This can be a video walkthrough or a document of screenshots that a novice user with little programming experience (the main audience for the shiny app) can follow along.

I will also just note that in so far as `app.R` already exists in `/inst`, there's no need to copy paste the app code into `shiny_mousebreedeR()`. That user-facing function can just be:

```r
shiny_mousebreedeR <- function() {
  appfile <- system.file("app.R", package = "mousebreedeR")
  shiny::runApp(appfile)
}
```

Please let me know once instructions for using the shiny app are available so I can go ahead and test its features.
