### 2024-05-27 82d2093
Update README.md
added photos and vignette for how to use shiny
