+++ README.md
@@ -13,29 +13,57 @@ MousebreedeR takes into account the exact genotypes the user has available in th
 
 The attached vignette walks the user through all of mousebreedeR's functionality and visualiztion recommendations to make the best possible choice to conduct the most efficient science possible. 
 
-# How to install
-You don't __need__ to install anything. You can use the GUI found at https://sportiellomike.shinyapps.io/mousebreedeR/. However, you can also download the package and vignette that comes with it as well.
+# How to actually use the software
+Generally, I'd recommend just using the graphical user interface (GUI) "shiny app" unless you have a good reason not to (https://sportiellomike.shinyapps.io/mousebreedeR/) since it can be used by true novices, and is the quickest way to get results. If you use this GUI at the link provided, you don't need to download anything.
 
-To install, go to Rstudio or the R command line and run the following code.
+## Using the shiny app (GUI)
 
-`install.packages('devtools')`
+* Go to any browser and go to the following URL: https://sportiellomike.shinyapps.io/mousebreedeR/
+* The outputs of the function will be found on the right side of the screen. The inputs required will be found on the left sidebar.
+![Picture of shiny app as it loads before inputs are put in](./README_images/Starting_screenshot_with_arrows.png)
 
-`library(devtools)`
+* Inputs
+	* Breeder mice CSV
+		* The inputs include both a CSV of breeder mice available formatted as one mouse per line, one column per gene of interest, and one column called "sex" on the right denoting "F" for female or "M" for male.
+	 	* You can have as many gene columns and mouse rows as you want.
+		* Each cell in the gene column should be 0, 1, or 2, representing the number of alleles in the mouse (0 for homozygous negative, 1 for heterozygous, or 2 for homozygous positive)
+	 	* Once your CSV is ready with your breeder mice you have available, upload your CSV by clicking where the read arrow points in the screenshot above. An example CSV is pictured below.
+	![Example breeder CSV](./README_images/CSV_formatting.png)
+	* Desired vec
+		* The software also requires the user to supply what their desired mouse is. For example, you may want a mouse to be a heterozygote at all three genes of interest.
+	 	* To supply your desired genotypes, create your desired vector (noted as desiredvec in this package) by writing down your desired genotype in the text field found in the sidebar (denoted by the blue arrow).
+   		* The options for each position in the vector are "homopos" for homozygous positive, "het" for heterozygote, or "homoneg" for homozygous negative. 
+   		* The correct format should read as follows: `het,het,het'.
+     		* The commas are necessary, and no need to add any quotes. The first position refers to the gene of interest that your leftmost column in your breedermouse CSV represents. The second position to the column to its right, the third position to the column to its right, etc.
+       * As soon as you upload both the CSV and type your desiredvec, the software will start working. No need to "hit go" or anything like that.
 
-`install_github('sportiellomike/mousebreedeR',build_vignettes = T)`
+* Outputs
+![Example breeder CSV](./README_images/All_Outputs.png)
+	* Outputs are pictured above. The top output is just the breeder CSV that the user supplied. This is to help the user make sure they formatted their starter CSV correctly.
+ 	* Under that is the desiredvec the user input to make sure that they input it correctly. The program should have put quotation marks around each comma separated value
+  	* The next section tells the user if the desired mouse can be achieved in only one cross, or if it will require more than one cross.
+  	* Next, advice is given by the software for which mice to pair together. If the mouse is not attainable in one cross, it will try to maximize the number of alleles (homozygous positive) at each position as that will increase the end probability of achieving both a heterozygote and a homozygous positive mouse.
[…]
