### 2024-05-29 #6474 @yjunechoe
Thanks - the package seems to be at a state where it meets all the minimum criteria for the checklist items. @csoneson all items have been checked off.

@sportiellomike In the meantime, please continue working to improve the results of `devtools::check()`, ideally to a point where no errors and warnings are detected. I personally haven't found any of them to be critical for installation on my end, but errors and warnings are generally indicative of a package's fragility across different user environments (for example, one such error suggests that you assume users to have `{DT}` installed due to the shiny app having a dependency on it - this should be declared in Imports). I recommend consulting the [Writing R Extensions](https://cran.r-project.org/doc/manuals/R-exts.html) manual and the [R Packages](https://r-pkgs.org/) book as you work through this.
