+++ paper.bib
@@ -8,8 +8,7 @@
 0013-7227 (Linking)},
    DOI = {10.1210/endocr/bqaa114},
    url = {https://www.ncbi.nlm.nih.gov/pubmed/32645712},
-   year = {2020},
-   type = {Journal Article}
+   year = {2020}
 }
 
 @article{RN203,
@@ -23,6 +22,5 @@
 1932-6203 (Linking)},
    DOI = {10.1371/journal.pone.0012418},
    url = {https://www.ncbi.nlm.nih.gov/pubmed/20811634},
-   year = {2010},
-   type = {Journal Article}
+   year = {2010}
 }
