+++ paper.md
@@ -6,18 +6,18 @@ tags:
   - experimental organism
   - colony management
 authors:
-  - name: Mike Sportiello PhD
+  - name: Mike Sportiello
     orcid: 0000-0003-1690-8702
     corresponding: true
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
-  - name: David J. Topham MS/PhD
+  - name: David J. Topham
     orcid: 0000-0002-9435-8673
     affiliation: 1
 
 affiliations:
- - name: Center for Vaccine Biology and Immunology, University of Rochester Medical Center, Rochester, NY 14642, USA
+ - name: Center for Vaccine Biology and Immunology, University of Rochester Medical Center, Rochester, NY 14642, United States of America
    index: 1
- - name: Medical Scientist Training Program, University of Rochester Medical Center, Rochester, NY 14642, USA
+ - name: Medical Scientist Training Program, University of Rochester Medical Center, Rochester, NY 14642, United States of America
    index: 2
 
 date: 28 December 2023
@@ -27,13 +27,13 @@ bibliography: paper.bib
 
 # Summary
 
-With the advent of gene editing playing a regular role to produce new experimental models, the need to get the appropriate allele at each respective locus in a model organism is now common. Instead of simple "knockout" vs "wildtype" mouse experiments, for example, it is common for there to be a gene of interest that is either floxed or not floxed depending on the experimental group, another gene to mark cells of a certain phenotype such as a fluorescent reporter, a cre expression cassette to flox the gene of interest out, and more. While it is often obvious how to efficiently breed experimental organisms to obtain one locus of interest, once three or more loci are involved this becomes difficult to assess. MousebreedeR is a free and open source command line R package that takes desired mouse genotypes as well as the genotypes of potential mouse breeders the user already has on hand as inputs, and delivers efficient breeding schema to the user. In addition, it supplies the user with the probability of producing each potential pup, the probability that no pups of interest are born of the desired genotype in a litter, and more. We have also supplied code that assists the user in plotting these tables to help them visualize crosses of interest to help them make decisions in their breeding.  
+With the advent of gene editing playing a regular role in producing new experimental models, the need to get the appropriate allele at each respective locus in a model organism is now common. Instead of simple "knockout" vs "wildtype" mouse experiments, for example, it is common for there to be a gene of interest that is either floxed or not floxed depending on the experimental group, another gene that marks cells of a certain phenotype such as a fluorescent reporter, a cre expression cassette to flox the gene of interest out, and more. While it is often obvious how to efficiently breed experimental organisms to obtain one locus of interest, once three or more loci are involved this becomes difficult to assess. MousebreedeR is a free and open source command line R package that takes desired mouse genotypes as well as the genotypes of potential mouse breeders that the user already has on hand as inputs, and delivers efficient breeding schema as the output. In addition, it supplies the user with the probability of producing each potential pup, the probability that no pups with the desired genotype are born in a litter, and more. We have also supplied code that assists the user in plotting these tables to help them visualize crosses of interest and make decisions in their breeding.  
 
 # Statement of need
 
[…]
