+++ paper.md
@@ -27,16 +27,16 @@ bibliography: paper.bib
 
 # Summary
 
-With the advent of gene editing playing a regular role in producing new experimental models, the need to get the appropriate allele at each respective locus in a model organism is now common. Instead of simple "knockout" vs "wildtype" mouse experiments, for example, it is common for there to be a gene of interest that is either floxed or not floxed depending on the experimental group, another gene that marks cells of a certain phenotype such as a fluorescent reporter, a cre expression cassette to flox the gene of interest out, and more. While it is often obvious how to efficiently breed experimental organisms to obtain one locus of interest, once three or more loci are involved this becomes difficult to assess. MousebreedeR is a free and open source command line R package that takes desired mouse genotypes as well as the genotypes of potential mouse breeders that the user already has on hand as inputs, and delivers efficient breeding schema as the output. In addition, it supplies the user with the probability of producing each potential pup, the probability that no pups with the desired genotype are born in a litter, and more. We have also supplied code that assists the user in plotting these tables to help them visualize crosses of interest and make decisions in their breeding.  
+With the advent of gene editing playing a regular role in producing new experimental models, the need to get the appropriate allele at each respective locus in a model organism is now common. Instead of simple "knockout" vs "wildtype" mouse experiments, for example, it is common for there to be a gene of interest that is either floxed (surrounded by known genetic sequence that allow for a gene to be knocked out when coexpression of the cre protein is expressed) or not floxed depending on the experimental group, another gene that marks cells of a certain phenotype such as a fluorescent reporter, a cre expression cassette to flox the gene of interest out, and more. While it is often obvious how to efficiently breed experimental organisms to obtain one locus of interest, once three or more loci are involved this becomes difficult to assess. MousebreedeR is a free and open source web-based shiny app and command line R package that takes desired mouse genotypes as well as the genotypes of potential mouse breeders that the user already has on hand as inputs, and delivers efficient breeding schema as the output. In addition, it supplies the user with the probability of producing each potential pup, the probability that no pups with the desired genotype are born in a litter, and more. We have also supplied code that assists the user in plotting these tables to help them visualize crosses of interest and make decisions in their breeding.  
 
 # Statement of need
 
-Efficient breeding is a scientific imperative: with a limited period of fertility in experimental organisms, users may only have limited attempts with organisms of desired genotypes [@RN204; @RN203]. Taking into account both gestational time and the necessary time it takes to wean pups off their mother's milk, an inefficient cross can set the user back months in the case of a female pregnant with the non-optimal male's offspring. Furthermore, as most research centers pay per diem costs, inefficiency is a large direct cost for many experimentalists. Individual cages often cost hundreds of dollars per year just in maintenance fees, not even accounting for the cost of extra rounds of genotyping (requiring PCR primers, PCR mastermix, gels, dye, etc) or for the time dedicated to the project by employees. Ethical and animal welfare concerns also arise as inefficient breeding results in the mass euthanasia of non-desired animals as well as ear/tail-clipping required for genotyping, which may cause animals pain and distress.
[…]
