### 2024-05-31 4694fda
Create CITATION.cff
