### 2024-05-31 561f95f
Delete CITATION.cff
