+++ CITATION.cff
@@ -1,12 +1,0 @@
-cff-version:  v1.0.0
-message: "If you use this software, please cite it as below."
-authors:
-- family-names: "Mike"
-  given-names: "Sportiello"
-  orcid: "https://orcid.org/0000-0003-1690-8702"
-- family-names: "David"
-  given-names: "Topham"
-  orcid: "https://orcid.org/0000-0002-9435-8673"
-title:
-version: v1.0.0
-date-released: 2024-05-31
