### 2024-05-31 b39a574
Update CITATION.cff
