### 2024-05-31 #6474 @sportiellomike
-     DONE: Double check authors and affiliations (including ORCIDs)
-     DONE (Version number is v1.0.0): Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
-     DONE (DOI: [10.5281/zenodo.11405146](https://zenodo.org/doi/10.5281/zenodo.11405146)): Archive the release on Zenodo/figshare/etc and post the DOI here.
-     DONE: Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
-     DONE: Make sure that the license listed for the archive is the same as the software license.
