### 2024-05-31 027d3b6
Delete CITATION.cff
