### 2024-06-02 #6474 @csoneson
Thanks. Could you please edit the metadata of the Zenodo archive so that the title agrees with that of the paper? Also the format of the last author's name is slightly different (with/without the middle 'J').

Also one thing that I noticed only now - the GitHub repo and the LICENSE.md file gives the license as MIT; however, the License field in the DESCRIPTION file says 'CC BY 4.0'. Could you please change the latter to also be MIT?
