### 2024-06-02 ca7d930
Update DESCRIPTION
appropriately changed license in description to MIT to match that of actual license file
