### 2024-06-02 #6474 @sportiellomike
- Title changed
- J appropriately added to author
- Description file now reflects correct license of MIT consistent with rest of package.

Sorry i missed those things @csoneson.
